+++ src/components/layouts/header/useHeader.ts
@@ -1,20 +1,12 @@
 import { signOutUser } from '@/lib/api';
-import { fetchCartItems } from '@/lib/api';
 import { userAtom } from '@/lib/store';
-import { useQuery } from '@tanstack/react-query';
 import { useAtom } from 'jotai';
 import { useRouter } from 'next/navigation';
 
 export default function useHeader() {
+  const router = useRouter();
   const [user, setUser] = useAtom(userAtom);
   const userEmail = user?.email ?? '';
-  const router = useRouter();
-  const { data: cartItems } = useQuery({
-    queryKey: ['cart'],
-    queryFn: async () => await fetchCartItems(user?.id ?? ''),
-  });
-
-  const badgeCount = cartItems?.length ?? 0;
 
   const handleLogout = async () => {
     await signOutUser();
@@ -22,14 +14,8 @@ export default function useHeader() {
     router.push('/');
   };
 
-  const handleCartButton = () => {
-    router.push('/cart');
-  };
-
   return {
-    badgeCount,
     userEmail,
     handleLogout,
-    handleCartButton,
   };
 }
